# contracts/backend/openapi.yaml
openapi: 3.1.0
info:
  title: DIP3 Java Analysis Backend
  version: 2.0.0
paths:
...
  /api/analysis/sessions/{sessionId}/follow-ups:
    get:
      summary: List the authenticated session follow-ups in creation order
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200':
          description: Session follow-ups
          content:
            application/json: { schema: { type: array, items: { $ref: './schemas/analysis-follow-up.schema.json' } } }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
    post:
      summary: Create a follow-up from a completed root or parent follow-up conclusion
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, minLength: 1 } }
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string, minLength: 1, maxLength: 300 }
                parentFollowUpId: { type: string, maxLength: 200 }
          multipart/form-data:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string, minLength: 1, maxLength: 300 }
                parentFollowUpId: { type: string, maxLength: 200 }
      responses:
        '303': { description: Created follow-up redirect with followUpId, headers: { Location: { schema: { type: string } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { description: The referenced execution has no persisted completed conclusion, content: { application/json: { schema: { $ref: './schemas/error.schema.json' } } } }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/analysis/sessions/{sessionId}/follow-ups/{followUpId}:
    get:
      summary: Return one authenticated session follow-up
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: followUpId, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200': { description: Follow-up detail, content: { application/json: { schema: { $ref: './schemas/analysis-follow-up.schema.json' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/analysis/sessions/{sessionId}/follow-ups/{followUpId}/context:
    post:
      summary: Apply an explicit context correction and invalidate an obsolete plan
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: followUpId, in: path, required: true, schema: { type: string, minLength: 1 } }
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded: { schema: { $ref: '#/components/schemas/FollowUpContextForm' } }
          multipart/form-data: { schema: { $ref: '#/components/schemas/FollowUpContextForm' } }
      responses:
        '303': { description: Updated, conflict-confirmation, or validation redirect, headers: { Location: { schema: { type: string } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/analysis/sessions/{sessionId}/follow-ups/{followUpId}/replan:
    post:
      summary: Build and persist a new governed follow-up plan version and diff
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: followUpId, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '303': { description: Replanned or explicit failure redirect, headers: { Location: { schema: { type: string } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
...
  /api/system/graph-sync/bootstrap:
    post:
      summary: Run the synchronous system-only full graph bootstrap
      description: Captures all seven ERP watermarks, rebuilds every active organization and advances all cursors only after complete success. This operation is intentionally not exposed through Next.js.
      security: [{ graphSyncOpsSecret: [] }]
      responses:
        '200': { description: Terminal full-bootstrap parent run, content: { application/json: { schema: { $ref: './schemas/graph-sync-shared.schema.json#/$defs/run' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/system/graph-sync/bootstrap/status:
    get:
      summary: Return the latest full graph bootstrap parent fact
      security: [{ graphSyncOpsSecret: [] }]
      responses:
        '200': { description: Latest full-bootstrap parent run, content: { application/json: { schema: { $ref: './schemas/graph-sync-shared.schema.json#/$defs/run' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
...
  /api/admin/graph-sync/status:
    get:
      summary: Return the authenticated organization graph-sync status
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: Scoped graph-sync status, content: { application/json: { schema: { $ref: './schemas/graph-sync-shared.schema.json#/$defs/status' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/admin/graph-sync/organizations/{organizationId}/status:
    get:
      summary: Return the latest run for an organization in the authenticated scope
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: organizationId, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200': { description: Latest organization run, content: { application/json: { schema: { $ref: './schemas/graph-sync-shared.schema.json#/$defs/run' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/admin/graph-sync/organizations/{organizationId}/rebuild:
    post:
      summary: Rebuild one authorized organization projection synchronously
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: organizationId, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200': { description: Terminal organization run, content: { application/json: { schema: { $ref: './schemas/graph-sync-shared.schema.json#/$defs/run' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/admin/graph-sync/consistency-sweep:
    post:
      summary: Run a scoped consistency rebuild as a platform administrator
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Completed organization runs
          content:
            application/json: { schema: { type: array, items: { $ref: './schemas/graph-sync-shared.schema.json#/$defs/run' } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: dip3_session }
    graphSyncOpsSecret: { type: apiKey, in: header, name: X-Graph-Sync-Ops-Secret }
  responses:
    Error:
      description: Explicit failure
      content:
        application/json: { schema: { $ref: './schemas/error.schema.json' } }
  schemas:
    FollowUpContextForm:
      type: object
      properties:
        targetMetric: { type: string, maxLength: 200 }
        entity: { type: string, maxLength: 200 }
        timeRange: { type: string, maxLength: 200 }
        comparison: { type: string, maxLength: 200 }
        factor: { type: string, maxLength: 200 }
        confirmConflicts: { type: string, enum: ['true', 'false'] }